+++ mono_repo/test/src/expected_output.dart
@@ -39,11 +39,11 @@ for PKG in ${PKGS}; do
       echo
       echo -e "\033[1mPKG: ${PKG}; TASK: ${TASK}\033[22m"
       case ${TASK} in
-      dartanalyzer)
+      analyze)
         echo 'dart analyze'
         dart analyze || EXIT_CODE=$?
         ;;
-      dartfmt)
+      format)
         echo 'dart format --output=none --set-exit-if-changed .'
         dart format --output=none --set-exit-if-changed . || EXIT_CODE=$?
         ;;